Make gcc14 the primary version on Tiger+
gcc7 is obsolete at this point and fails to build a lot of software.
Enables gcc14 for all platforms as sets it as default.

gcc14 is built using gcc10-bootstrap on 10.4 and 10.5.

This does not enable other versions of gcc between 9 and 14
because there doesn't really seem to be a reason to use those
given that gcc14 is available (this can be revisited later if turns out
that someone needs those other versions).